scraper/requirements.txt
flask
flask-cors
celery
redis
requests
beautifulsoup4
psycopg2-binary